requirements.txt

Pillow
customtkinter
gui
logic
requests
